Fix Confusion + Mirror Move + Speed Tie interaction
Fuzz tests found a scenario where (+confusion) is possible from a
(-confusion) template: if a Pokemon using Mirror Move speed ties its
confused opponent that is using a move that causes confusion then we
need to also account for the possibility confusion could ended and
then be started. This is incredibly niche and also dependent on the
seed state as usually the transitions algorithm won't get stuck.
This should also only be relevant for confusion and not the other
durations as all the other durations should prevent one from using a
move (can't use a sleep move while asleep etc). This fix which
iterates over additional possibilities will obviously bloat the
number of visited states and is thus scoped to only happen in speed
tie scenarios