fix overflow from gen 1 transitions
the intermediate calculation 41956097078536962048 * 20978048539268481024
requires more than 128 bits to store, meaning Rational(u128) can't
work. Rational(u256) doesn't work because u256 isn't allowed in an
extern struct, so we're forced to switch to BigRational. Thankfully
this should only be required for internal bookkeeping (intermediate
debugging output and for checking things sum to 1) - most operations
can still be done on a u128.