allow for larger fixed-width Rationals
switching to BigRational in 26a79a0 resulted in a fairly large
performance regression, using a Rational(u256) mitigates most of
this (6-7x faster).

Unfortunately, the u256 field trips an issue with Zig's compiler_rt
support: ziglang/zig#22667. Not the end of the world given its only
in the debugging code so it can just be commented out temporarily